Breakup Page type and member functions to page package
This includes a refactor to generateAssets to remove some duplication
for copying files and making new directories.
